Move FakeFieldContainer to detection.py as FieldDetector.
