+++ ytree/data_structures/detection.py
@@ -20,6 +20,27 @@ from ytree.utilities.exceptions import ArborFieldNotFound
 
 _selectors = ("forest", "tree", "prog")
 
+class FieldDetector(defaultdict):
+    """
+    A fake field data container used to calculate dependencies.
+    """
+    def __init__(self, arbor, name=None):
+        self.arbor = arbor
+        self.name = name
+
+    def __missing__(self, key):
+        if key not in self.arbor.field_info:
+            raise ArborFieldDependencyNotFound(
+                self.name, key, self.arbor)
+        fi = self.arbor.field_info[key]
+        units = fi.get("units", "")
+        if fi.get("vector_field", False):
+            data = np.ones((1, 3))
+        else:
+            data = np.ones(1)
+        self[key] = self.arbor.arr(data, units)
+        return self[key]
+
 class SelectionDetector(defaultdict):
     """
     A TreeNode-like object to test select_halos criteria.
